Changed how updates are handled.
If the server and client were both restarted and the server had no
record of an existing contact that the client had. If the client then
updated that record it would not exist in the server to update and
would result in no record in the server.
Changed the logic in the server to send updates to the log_contact
method since it first checks if the unique id exists before it either
logs as new, or updates an existing.